updated delete method in posts and users controller to delete pictures that belong to the post or user from photos database

+++ app/Http/Controllers/AdminPostsController.php
@@ -129,7 +129,10 @@ class AdminPostsController extends Controller
     {
         //
         $record = Post::findOrFail($id);
-        unlink(public_path().$record->photo->path);
+        if($record->photo) {
+            unlink(public_path().$record->photo->path);
+            $record->photo()->delete();
+        };
         $record->delete();
         Session::flash('deleted_post','The Post has been deleted');
         return redirect(route('admin.posts.index')) ;
